# backend/internal/api/go/model_expense_response.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Ledger API
 *
 * This is the API for the ledger backend.
 *
 * API version: 0.0.1
 */

package openapi




type ExpenseResponse struct {

	Id string `json:"id"`
}

// AssertExpenseResponseRequired checks if the required fields are not zero-ed
func AssertExpenseResponseRequired(obj ExpenseResponse) error {
	elements := map[string]interface{}{
		"id": obj.Id,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertExpenseResponseConstraints checks if the values respects the defined constraints
func AssertExpenseResponseConstraints(obj ExpenseResponse) error {
	return nil
}