# backend/internal/api/go/model_income.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Ledger API
 *
 * This is the API for the ledger backend.
 *
 * API version: 0.0.1
 */

package openapi


import (
	"errors"
)



type Income struct {

	Name string `json:"name"`

	Amount float32 `json:"amount"`

	Date string `json:"date"`

	Received bool `json:"received"`
}

// AssertIncomeRequired checks if the required fields are not zero-ed
func AssertIncomeRequired(obj Income) error {
	elements := map[string]interface{}{
		"name": obj.Name,
		"amount": obj.Amount,
		"date": obj.Date,
		"received": obj.Received,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertIncomeConstraints checks if the values respects the defined constraints
func AssertIncomeConstraints(obj Income) error {
	if obj.Amount < 0 {
		return &ParsingError{Param: "Amount", Err: errors.New(errMsgMinValueConstraint)}
	}
	return nil
}