# backend/internal/api/go/model_income_response.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Ledger API
 *
 * This is the API for the ledger backend.
 *
 * API version: 0.0.1
 */

package openapi




type IncomeResponse struct {

	Id string `json:"id"`
}

// AssertIncomeResponseRequired checks if the required fields are not zero-ed
func AssertIncomeResponseRequired(obj IncomeResponse) error {
	elements := map[string]interface{}{
		"id": obj.Id,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertIncomeResponseConstraints checks if the values respects the defined constraints
func AssertIncomeResponseConstraints(obj IncomeResponse) error {
	return nil
}